requirements.txt
streamlit>=1.29.0
python-dotenv>=1.0.0
openai>=1.6.1
autogen-agentchat>=0.2.0
autogen-core>=0.2.0
autogen-ext>=0.2.0
requests>=2.31.0
pandas>=2.1.4
numpy>=1.26.2
matplotlib>=3.8.2
yfinance>=0.2.33
scipy>=1.11.4 